+++ src/components/stories/Components.stories.tsx
@@ -14,7 +14,7 @@ import InlineWidget from "../InlineWidget/InlineWidget";
 import PopupButton from "../PopupButton/PopupButton";
 import PopupWidget from "../PopupWidget/PopupWidget";
 import CalendlyEventListener from "../CalendlyEventListener/CalendlyEventListener";
-import { PageSettings, Utm, Prefill, openPopupWidget } from "../../calendly";
+import { PageSettings, Utm, Prefill } from "../../calendly";
 
 const prefill: Prefill = {
   name: "Jon Snow",
@@ -34,7 +34,7 @@ const prefill: Prefill = {
     a9: "a9",
     a10: "a10",
   },
-  date: new Date(Date.now() + 86400000)
+  date: new Date(Date.now() + 86400000),
 };
 
 const utm: Utm = {
@@ -79,6 +79,7 @@ storiesOf("Components", module)
       styles={object("styles", {})}
       className={text("className", "")}
       iframeTitle={text("iframeTitle", "Calendly Scheduling Page")}
+      rootElement={document.getElementById("root")!}
     />
   ))
   .add("PopupWidget", () => (
@@ -92,33 +93,9 @@ storiesOf("Components", module)
       utm={object("utm", utm)}
       pageSettings={object("pageSettings", pageSettings)}
       iframeTitle={text("iframeTitle", "Calendly Scheduling Page")}
+      rootElement={document.getElementById("root")!}
     />
   ))
-  .add("Custom Button", () => {
-    const options = {
-      url: text("url", "https://calendly.com/acmesales"),
-      pageSettings: object("pageSettings", pageSettings),
-      utm: object("utm", utm),
-      prefill: object("prefill", prefill),
-      iframeTitle: text("iframeTitle", "Calendly Scheduling Page"),
-    };
-
-    return (
-      <div>
-        <h4 style={{ textAlign: "center" }}>
-          {" "}
-          Use the <code>openPopupWidget</code> function to create a custom
-          button that will open the pop-up scheduler when clicked.
-        </h4>
-        <button
-          style={{ display: "block", margin: "0 auto" }}
-          onClick={() => openPopupWidget(options)}
-        >
-          Custom Button
-        </button>
-      </div>
-    );
-  })
   .add("CalendlyEventListener", () => {
     const eventId = "calendly-event";
     const instructions =
@@ -167,7 +144,7 @@ storiesOf("Components", module)
               frameBorder="0"
               height="100%"
               width="100%"
-              src={`https://calendly.com/acmesales?embed_domain=${document.location.host}&embed_type=Inline`}
+              src={`https://calendly.com/acmesales?embed_domain=1&embed_type=Inline`}
             />
           </div>
         </CalendlyEventListener>